Add files via upload
This script was made to classify negative response values as zero and positive values as 1 within the training data set for the Jane Street Market Prediction Competition. 

The purpose of this is to create a csv file that can be analyzed within Minitab as Minitab makes EDA a very easy process.