Create DecisionTreeExplore.py
This script was used to identify optimal training parameters for a decision tree. 

I wanted to see if optimal parameters for a decision tree would remain as optimal parameters for a random forest. My thought process here is that it took me less than 1 minute to train a decision tree and it took me over 4 hours to train a random forest so if I could identify optimal parameters for a tree that would also be optimal for a random forest I could save a lot of time. 

This theory was tried and proven to be false.